+++ src/blocking/mod.rs
@@ -58,7 +58,7 @@ use semver::Version;
 pub fn get_versions(crate_name: &str, user_agent: &str) -> Result<Versions> {
     let url = build_url(crate_name);
     let response: CratesioResponse = reqwest::blocking::Client::builder()
-        .user_agent(format!("{}/{}", crate_name, user_agent))
+        .user_agent(user_agent)
         .build()
         .context("Couldn't build client")?
         .get(&url)
@@ -73,7 +73,7 @@ pub fn get_versions(crate_name: &str, user_agent: &str) -> Result<Versions> {
 fn get_version_list(crate_name: &str, user_agent: &str) -> Result<Vec<Version>> {
     let url = build_url(crate_name);
     let response: CratesioResponse = reqwest::blocking::Client::builder()
-        .user_agent(format!("{}/{}", crate_name, user_agent))
+        .user_agent(user_agent)
         .build()
         .context("Couldn't build client")?
         .get(&url)