+++ src/blocking/mod.rs
@@ -84,6 +84,7 @@ fn get_version_list(crate_name: &str, user_agent: &str) -> Result<Vec<Version>> 
     let versions = response.all_versions;
     let versions = versions
         .into_iter()
+        .filter(|v| !v.yanked)
         .map(|v| v.version)
         .collect();
     Ok(versions)